MedVedev-AppV2/backend/requirements.txt
# Optional: load API keys from backend/.env
python-dotenv>=1.0.0

# PDF report (SOAP JSON → HTML → PDF)
jinja2>=3.1.0
weasyprint>=62.0

# API server
fastapi>=0.110.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Password hashing
passlib>=1.7.0

# JWT auth
python-jose[cryptography]>=3.5.0
